Added function getbysaleid for have a list of saleitem for the sale

+++ src/main/java/Managment/WareHouse/Repository/SalesRepository.java
@@ -3,5 +3,5 @@ package Managment.WareHouse.Repository;
 import Managment.WareHouse.Entity.Sales;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SalesRepository extends JpaRepository<Sales, Integer> {
+public interface SalesRepository extends JpaRepository<Sales, Long> {
 }
